requirements.txt
requests
Pillow
rtree
SlackClient
flask